Rename some classes in ProcessWindows.
Renamed monitor -> driver, to make clear that the implementation here
is in no way related to that of other process plugins which have also
implemented classes with similar names such as DebugMonitor.

Also created a DebugEventHandler interface, which will be used by
implementors to get notified when debugging events happen in the
inferiors.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@221322 91177308-0d34-0410-b5e6-96231b3b80d8